dvdplayer: move pts output queue into DVDAudio
